Refactoring my code
Changing structure of models and fixing some errors in code. The main goal is to do the structure of database in the way, that there will be stored only URI adress (file path) to image, but not image itself. So next step will be creating a method to save files in image folder, when user is registrated. And after that I will change database structure, and continue working on this project.